+++ src/__tests__/components/HomePage.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import HomePage from '../components/HomePage';
+import HomePage from '../../components/HomePage';
 import renderer from 'react-test-renderer';
 
 describe('<HomePage />', () => {
